AP-1350 Results can be grouped by field.
git-svn-id: https://opensource.kantega.no/svn/openaksess/branches/openaksess-solr@3637 fd808399-8219-4f14-9d4c-37719d9ec93d

+++ search-api/src/java/no/kantega/search/api/search/QueryStringGenerator.java
@@ -62,10 +62,10 @@ public class QueryStringGenerator {
         int startPage = ((page / 10) * 10) + 1;
         int endPage = startPage + 9;
 
-        long numberOfHits = searchResponse.getNumberOfHits();
+        int numberOfHits = searchResponse.getNumberOfHits().intValue();
         int resultsPerPage = query.getResultsPerPage();
         if (endPage * resultsPerPage >= numberOfHits) {
-            endPage = (int) ((numberOfHits - 1) / numberOfHits);
+            endPage = ((numberOfHits - 1) / numberOfHits);
             endPage++;
         }
         if (startPage > 1) {
